Split simulation modules into separate files
